関数signupに関するindex.jsのコードをuser.jsに移動
