+++ store/index.js
@@ -13,36 +13,6 @@ export const mutations = {
 };
 
 export const actions = {
-  // ユーザのアイコン画像をアップロード
-  async upLoadUserImage(ctx, file) {
-    await firebase.storage().ref().child(`images/${file.name}`).put(file)
-  },
-  // ユーザのアイコン画像のダウンロードURLを取得
-  async downLoadUserImage(ctx, file) {
-    return await firebase.storage().ref().child(`images/${file.name}`).getDownloadURL();
-  },
-  async setUserData(ctx, {userName, userImage}) {
-    const userData = firebase.auth().currentUser;
-    await userData.updateProfile({
-      displayName: userName,
-      photoURL: userImage
-    });
-  },
-  async createAccount(
-    {dispatch},
-    {userMail,
-    password, 
-    userName,
-    userImage}) {
-    await firebase.auth().
-          createUserWithEmailAndPassword(userMail, password);
-    await dispatch('setUserData', {userName, userImage});
-  },
-  showUserNameOnAlert(ctx) {
-    const userData = firebase.auth().currentUser;
-    console.log(userData);
-    alert(`こんにちは、${userData.displayName}さん！登録完了です！` );
-  },
   async storageData(
     { state },
     {movieTitle,
